refactor(tui): split dashboard/model.go into focused files
The single 685-line model.go combined Model definition, message
types, Update/View/render logic, and visual-width padding
helpers. Split into focused siblings within the same package:

  model.go      -- Model + lifecycle hooks
  update.go     -- Update reducer
  helpers.go    -- toast/levelOf/opResultText/startGHAuth/hasRateLimitedAccounts
  messages.go   -- bubbletea message type definitions
  view.go       -- View, renderHelp
  render.go     -- renderAccountTable
  pad.go        -- padPlain, padStyled, ansiPad, min

Pure file-level reorganization. No symbols renamed. No behavior
change. PR #242 model_test.go continues to pass since all moves
stay within package dashboard.

Closes #249
